requirements.txt
pybooru>=4.2.2
python-dotenv>=0.19.0 